Update test packages and refactor unit test
Several test-related packages have been updated to their latest versions across multiple projects. Furthermore, a unit test was moved from the integration tests to the appropriate unit tests class, ensuring proper organization and categorization of tests.
