+++ test/ShapeCrawler.Tests.Integration/PresentationITests.cs
@@ -2,7 +2,6 @@ using System;
 using System.IO;
 using System.Text.Json;
 using FluentAssertions;
-using NUnit.Framework;
 using ShapeCrawler.Logger;
 using ShapeCrawler.Tests.Shared;
 using ShapeCrawler.Tests.Unit.Helpers;
@@ -112,31 +111,6 @@ public class PresentationITests : SCTest
         autoShapeText.Should().BeEquivalentTo(originalText);
             
         // Clean
-        File.Delete(newPath);
-    }
-    
-    [Test]
-    public void SaveAs_should_not_change_the_Original_Path_when_it_is_saved_to_New_Path()
-    {
-        // Arrange
-        var originalPath = GetTestPath("001.pptx");
-        var pres = new Presentation(originalPath);
-        var textFrame = pres.Slides[0].Shapes.GetByName<IShape>("TextBox 3").TextFrame;
-        var originalText = textFrame!.Text;
-        var newPath = Path.GetTempFileName();
-        textFrame.Text = originalText + "modified";
-
-        // Act
-        pres.SaveAs(newPath);
-
-        // Assert
-        pres = new Presentation(originalPath);
-        textFrame = pres.Slides[0].Shapes.GetByName<IShape>("TextBox 3").TextFrame;
-        var autoShapeText = textFrame!.Text; 
-        autoShapeText.Should().BeEquivalentTo(originalText);
-            
-        // Clean
-        File.Delete(originalPath);
         File.Delete(newPath);
     }
 }
