+++ test/ShapeCrawler.Tests.Shared/ShapeCrawler.Tests.Shared.csproj
@@ -20,7 +20,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="NUnit" Version="3.13.3" />
+    <PackageReference Include="NUnit" Version="4.0.1" />
   </ItemGroup>
 
 </Project>
